requirements.txt
aio-pika
fastapi
python-multipart
uvicorn
pydantic
alembic
sqlalchemy
sqlalchemy-utils
dynaconf[yaml]==3.1.4
psycopg2-binary==2.8.6
databases
asyncpg
pyjwt
starlette==0.13.8
passlib